ERR: change meta output from double to float since meta is currently broken for byte swapping double.

+++ Testing/Code/Algorithms/itkSTAPLEImageFilterTest.cxx
@@ -53,7 +53,7 @@ template< unsigned int VDimension>
 class Stapler : public StaplerBase
 {
 public:
-  typedef itk::Image< double, VDimension > OutputImageType;
+  typedef itk::Image< float, VDimension > OutputImageType;
   typedef itk::Image< unsigned short, VDimension > InputImageType;
   typedef itk::STAPLEImageFilter<InputImageType, OutputImageType> StapleFilterType;
